+++ tests/testthat/test_MSnExp.R
@@ -375,7 +375,7 @@ test_that("pData<- on MSnExp works", {
 test_that("injection time", {
     f <- msdata::proteomics(full.names = TRUE, pattern = "MS3")
     it1 <- MSnbase:::injectionTimeFromFile1(f) ## in millisecs
-    it2 <- fData(readMSData2(f))$injectionTime ## in secs
+    it2 <- fData(readMSData(f, mode = "onDisk"))$injectionTime ## in secs
     expect_equal(it1/1000, it2)
 })