+++ tests/testthat/test_id.R
@@ -54,7 +54,7 @@ test_that("adding compatible ident to MSnExp and MSnSet", {
                       full.name = TRUE, pattern = "mzXML$")
      identFile <- dir(system.file(package = "MSnbase", dir = "extdata"),
                       full.name = TRUE, pattern = "dummyiTRAQ.mzid")
-     msexp <- readMSData2(quantFile)
+     msexp <- readMSData(quantFile, mode = "onDisk")
      msset <- quantify(msexp, method = "max", reporters = iTRAQ4)
      msexp <- addIdentificationData(msexp, identFile)     
      msset <- addIdentificationData(msset, identFile)